Add tests and changes for rendering table to csv
Change row._get_value to accept column name instead of a bound column.
Add unit tests for rendering the table to csv with various options.
Make sure column visibility is honored when rendering the table
to csv format.
Remove dependency on unicodecsv, however, allow it to be the default csv
module if installed.

Conflicts:

	tests/templates.py